mpegmap/HTMLSourcingInbandTracks-3.3-4.html

<!doctype html>
<html>
  <head>
    <title>TC 5.2-1 Closed Caption TextTrack</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, Closed Caption - UA creates TextTrack, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing both 608 and 708 Closed Caption.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>

    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>

    <script>
      var test = async_test("Closed Caption TextTrack attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "ClosedCaption.ts";

      var serviceNumber = '1';  // service number in service block
      var id = 'sn1';  // caption service number in Caption Service Descriptor
      var kind = 'captions';
      var language = 'en';

      vid.textTracks.addEventListener("addtrack", function(event) {
        var track = event.track;
        track.mode = "hidden";
        if (track.id === id) {
          test.step(function() {
            assert_equals(track.kind, kind, "TextTrack.kind");
            assert_equals(track.language, language, "TextTrack.language");
          });
          test.done();
        }
      });
      vid.addEventListener("playing", function() {
        vid.pause();
        if (!vid.textTracks.getTrackById(id)) {
          test.step(function() {
            assert_unreached("Did not find Closed Caption Track with id " +
                             id + " in DOM.");
          });
          test.done();
        }
      });
    </script>
  </body>
</html>
